Remove IO functions (moved to loaders.gl)
