+++ modules/test-utils/src/index.js
@@ -1,6 +1,5 @@
 export {toLowPrecision} from './utils/precision';
 export {default as gl} from './utils/setup-gl';
-export * from './luma.gl/gpgpu';
 
 // Utilities for update tests (lifecycle tests)
 export {testLayer} from './lifecycle-test';
