--- modules/test-utils/src/luma.gl/gpgpu/index.js
@@ -1,3 +1,0 @@
-// GPGPU utilities for luma.gl
-
-export {diffImagePixels} from './diff-images';
