--- modules/test-utils/src/luma.gl/io-basic/browser-request-image.js
@@ -1,24 +1,0 @@
-import {getPathPrefix} from './path-prefix';
-/* global Image */
-
-/*
- * Loads images asynchronously
- * image.crossOrigin can be set via opts.crossOrigin, default to 'anonymous'
- * returns a promise tracking the load
- */
-export function loadImage(url, opts) {
-  const pathPrefix = getPathPrefix();
-  url = pathPrefix ? pathPrefix + url : url;
-
-  return new Promise((resolve, reject) => {
-    try {
-      const image = new Image();
-      image.onload = () => resolve(image);
-      image.onerror = () => reject(new Error(`Could not load image ${url}.`));
-      image.crossOrigin = (opts && opts.crossOrigin) || 'anonymous';
-      image.src = url;
-    } catch (error) {
-      reject(error);
-    }
-  });
-}
