--- modules/test-utils/src/luma.gl/io-basic/node-load-image.js
@@ -1,10 +1,0 @@
-import {promisify} from '../../utils';
-
-export function loadImage(url) {
-  const getPixels = module.require('get-pixels');
-  if (!getPixels) {
-    throw new Error('loadImage: get-pixels not installed');
-  }
-
-  return promisify(getPixels)(url);
-}
