--- modules/test-utils/src/luma.gl/io-basic/path-prefix.js
@@ -1,12 +1,0 @@
-let pathPrefix = '';
-
-/*
- * Set a relative path prefix
- */
-export function setPathPrefix(prefix) {
-  pathPrefix = prefix;
-}
-
-export function getPathPrefix() {
-  return pathPrefix;
-}
