--- modules/test-utils/src/luma.gl/io-extended/README.md
@@ -1,8 +1,0 @@
-# extended-io
-
-A separate module that contains IO functions that requires dependencies that can signficantly increase the size of an application bundle.
-
-More advanced IO support that handles e.g.
-* stream support in browsers
-
-is being developed in this separate module.
