--- modules/test-utils/src/luma.gl/io-extended/index.js
@@ -1,6 +1,0 @@
-export * from './browser-image-io';
-
-export {loadFile} from './browser-request';
-
-import * as fs from './browser-fs';
-export {fs};
