--- modules/test-utils/src/luma.gl/io-extended/test/index.js
@@ -1,1 +1,0 @@
-import './write-read-image.spec.js';
