--- modules/test-utils/src/luma.gl/test/index.js
@@ -1,1 +1,0 @@
-import './read-image.spec.js';
