+++ modules/test-utils/src/utils/deprecated/browser-image-utils.js
