+++ modules/test-utils/src/utils/deprecated/diff-images.js
