+++ modules/test-utils/src/utils/setup-gl.js
@@ -1,7 +1,7 @@
 import {setContextDefaults} from 'luma.gl';
 import {createGLContext} from 'luma.gl';
-import createContext from '../luma.gl/headless/headless';
-import {global} from '../luma.gl/utils/globals';
+// import createContext from '../luma.gl/headless/headless';
+import {global} from './deprecated/globals';
 
 setContextDefaults({
   width: 1,
@@ -11,11 +11,9 @@ setContextDefaults({
   // throwOnError: false
 });
 
-global.glContext =
-  global.glContext ||
-  createGLContext() ||
-  // TODO - Seems to be an issue in luma.gl
-  (createContext && createContext(100, 100, {}));
-// console.log('Context', global.glContext);
+global.glContext = global.glContext || createGLContext();
+//   // TODO - Seems to be an issue in luma.gl
+//   (createContext && createContext(100, 100, {}));
+// // console.log('Context', global.glContext);
 
 export default global.glContext;
